COnfig two types verperf and cdn

+++ vulkan/src_app/particleOnly/ResourceAtomicCompute.cpp
@@ -152,7 +152,7 @@ void ResourceAtomicCompute::AskObject(uint32_t AnyNumber)
 	//m_App->m_Numparticles = m_collisionStruct.numParticles;
 	m_PerfObj->m_ReportBuffer[AnyNumber].NumCollisionsComputeCount = m_collisionStruct.CollisionCount;
 	m_PerfObj->m_ReportBuffer[AnyNumber].NumParticlesComputeCount = uint32_t(m_collisionStruct.numParticles);
-	m_PerfObj->m_ReportBuffer[AnyNumber].ThreadCountComp = m_collisionStruct.holdPidx;
+	m_PerfObj->m_ReportBuffer[AnyNumber].ThreadCountComp =  uint32_t(m_collisionStruct.holdPidx);
 
 
 }
